requirements.txt
click==6.6
Flask==0.11.1
Flask-API==0.6.7
Flask-MySQL==1.4.0
Flask-MySQLdb==0.2.0
Flask-Script==2.0.5
itsdangerous==0.24
Jinja2==2.8
Markdown==2.6.7
MarkupSafe==0.23
mysqlclient==1.3.9
PyMySQL==0.7.9
Werkzeug==0.11.11
wheel==0.24.0
